VEP scoring: cast scores to fp32 before numpy (bf16 backbones)
score_transition_pairs did loss.cpu().numpy(), which fails on bf16 tensors —
ESM-C / flash backbones run in bf16 and numpy has no bfloat16. Cast with .float().

+++ protevo/vep/_scoring.py
@@ -83,6 +83,7 @@ def score_transition_pairs(
             reduction="none",
         ).mean(dim=-1)
 
-        scores.extend(loss.cpu().numpy())
+        # .float(): scores are bf16 for flash/ESM-C backbones, and numpy has no bfloat16.
+        scores.extend(loss.float().cpu().numpy())
 
     return scores
